Move aggregation method to utils
